Splits status from main store
