+++ frontend/app/src/components/defi/Borrowing.vue
@@ -103,7 +103,7 @@ import StatCardWide from '@/components/display/StatCardWide.vue';
 import DefiProtocolSelector from '@/components/helper/DefiProtocolSelector.vue';
 import ProgressScreen from '@/components/helper/ProgressScreen.vue';
 import RefreshHeader from '@/components/helper/RefreshHeader.vue';
-import { setupStatusChecking, useRoute } from '@/composables/common';
+import { useSectionLoading, useRoute } from '@/composables/common';
 import { Section } from '@/store/const';
 import { useDefiSupportedProtocolsStore } from '@/store/defi/protocols';
 import { Module } from '@/types/modules';
@@ -118,7 +118,7 @@ const store = useDefiSupportedProtocolsStore();
 const route = useRoute();
 const { tc } = useI18n();
 
-const { shouldShowLoadingScreen, isSectionRefreshing } = setupStatusChecking();
+const { shouldShowLoadingScreen, isSectionRefreshing } = useSectionLoading();
 
 const loading = shouldShowLoadingScreen(Section.DEFI_BORROWING);
