+++ frontend/app/src/components/defi/Deposits.vue
@@ -111,7 +111,7 @@ import BlockchainAccountSelector from '@/components/helper/BlockchainAccountSele
 import DefiProtocolSelector from '@/components/helper/DefiProtocolSelector.vue';
 import ProgressScreen from '@/components/helper/ProgressScreen.vue';
 import RefreshHeader from '@/components/helper/RefreshHeader.vue';
-import { setupStatusChecking, useRoute } from '@/composables/common';
+import { useSectionLoading, useRoute } from '@/composables/common';
 import { usePremium } from '@/composables/premium';
 import { useInterop } from '@/electron-interop';
 import {
@@ -147,7 +147,7 @@ const protocol = ref<DefiProtocol | null>(null);
 const premium = usePremium();
 const route = useRoute();
 const { openUrl } = useInterop();
-const { shouldShowLoadingScreen, isSectionRefreshing } = setupStatusChecking();
+const { shouldShowLoadingScreen, isSectionRefreshing } = useSectionLoading();
 const { floatingPrecision } = storeToRefs(useGeneralSettingsStore());
 
 const defiStore = useDefiStore();
