+++ frontend/app/src/store/balances/manual.ts
@@ -17,9 +17,9 @@ import { Section, Status } from '@/store/const';
 import { useMessageStore } from '@/store/message';
 import { useNotifications } from '@/store/notifications';
 import { useGeneralSettingsStore } from '@/store/settings/general';
+import { getStatus, setStatus } from '@/store/status';
 import { useTasks } from '@/store/tasks';
 import { ActionStatus } from '@/store/types';
-import { getStatus, setStatus } from '@/store/utils';
 import { TaskMeta } from '@/types/task';
 import { TaskType } from '@/types/task-type';
 import { assert } from '@/utils/assertions';
