+++ frontend/app/src/store/defi/liquity/index.ts
@@ -9,11 +9,12 @@ import { acceptHMRUpdate, defineStore } from 'pinia';
 import { Ref, ref } from 'vue';
 import { usePremium } from '@/composables/premium';
 import { useModules } from '@/composables/session';
+import { useStatusUpdater } from '@/composables/status';
 import i18n from '@/i18n';
 import { api } from '@/services/rotkehlchen-api';
 import { Section } from '@/store/const';
+import { fetchDataAsync } from '@/store/fetch-async';
 import { OnError } from '@/store/typing';
-import { fetchDataAsync, getStatusUpdater } from '@/store/utils';
 import { Module } from '@/types/modules';
 import { TaskMeta } from '@/types/task';
 import { TaskType } from '@/types/task-type';
@@ -196,7 +197,7 @@ export const useLiquityStore = defineStore('defi/liquity', () => {
   }
 
   const reset = () => {
-    const { resetStatus } = getStatusUpdater(Section.DEFI_LIQUITY_BALANCES);
+    const { resetStatus } = useStatusUpdater(Section.DEFI_LIQUITY_BALANCES);
 
     set(balances, {});
     set(events, {});
