+++ frontend/app/src/store/defi/sushiswap/index.ts
@@ -4,6 +4,7 @@ import { acceptHMRUpdate, defineStore } from 'pinia';
 import { computed, Ref, ref } from 'vue';
 import { usePremium } from '@/composables/premium';
 import { useModules } from '@/composables/session';
+import { useStatusUpdater } from '@/composables/status';
 import i18n from '@/i18n';
 import { api } from '@/services/rotkehlchen-api';
 import { useAssetInfoRetrieval } from '@/store/assets/retrieval';
@@ -18,8 +19,8 @@ import {
   getPoolProfit,
   getPools
 } from '@/store/defi/xswap-utils';
+import { fetchDataAsync } from '@/store/fetch-async';
 import { OnError } from '@/store/typing';
-import { fetchDataAsync, getStatusUpdater } from '@/store/utils';
 import { Module } from '@/types/modules';
 import { TaskMeta } from '@/types/task';
 import { TaskType } from '@/types/task-type';
@@ -131,7 +132,7 @@ export const useSushiswapStore = defineStore('defi/sushiswap', () => {
   }
 
   async function reset() {
-    const { resetStatus } = getStatusUpdater(Section.DEFI_SUSHISWAP_BALANCES);
+    const { resetStatus } = useStatusUpdater(Section.DEFI_SUSHISWAP_BALANCES);
     set(balances, {});
     set(events, {});
     resetStatus();
