+++ frontend/app/src/store/defi/yearn/index.ts
@@ -3,6 +3,7 @@ import { acceptHMRUpdate, defineStore } from 'pinia';
 import { computed, ComputedRef, Ref, ref } from 'vue';
 import { usePremium } from '@/composables/premium';
 import { useModules } from '@/composables/session';
+import { useStatusUpdater } from '@/composables/status';
 import i18n from '@/i18n';
 import { balanceKeys } from '@/services/consts';
 import { ProtocolVersion } from '@/services/defi/consts';
@@ -18,13 +19,9 @@ import { api } from '@/services/rotkehlchen-api';
 import { useAssetInfoRetrieval } from '@/store/assets/retrieval';
 import { Section, Status } from '@/store/const';
 import { useNotifications } from '@/store/notifications';
+import { getStatus, setStatus } from '@/store/status';
 import { useTasks } from '@/store/tasks';
-import {
-  getStatus,
-  getStatusUpdater,
-  isLoading,
-  setStatus
-} from '@/store/utils';
+import { isLoading } from '@/store/utils';
 import { Module } from '@/types/modules';
 import { TaskMeta } from '@/types/task';
 import { TaskType } from '@/types/task-type';
@@ -309,7 +306,7 @@ export const useYearnStore = defineStore('defi/yearn', () => {
   }
 
   const reset = (version?: ProtocolVersion) => {
-    const { resetStatus } = getStatusUpdater(
+    const { resetStatus } = useStatusUpdater(
       Section.DEFI_YEARN_VAULTS_BALANCES
     );
     if (!version || version === ProtocolVersion.V1) {
